# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Artifex API",
        "description": "二维游戏美术协作与 AI 资产生成平台 API",
        "version": "1.4.0",
        "contact": { "name": "Artifex Team" }
    },
    "servers": [
        { "url": "http://127.0.0.1:3000", "description": "本地开发" }
    ],
...
    "paths": {
...
        "/api/asset-library": {
            "get": {
                "tags": ["asset-library"],
                "summary": "获取用户素材库",
                "parameters": [
                    { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "页码" },
                    { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }, "description": "每页条数" }
                ],
                "responses": { "200": { "description": "素材列表" } }
            },
            "post": {
                "tags": ["asset-library"],
                "summary": "添加素材到素材库",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name", "content"],
                                "properties": {
                                    "name": { "type": "string", "description": "素材名称" },
                                    "type": { "type": "string", "description": "素材类型" },
                                    "content": { "type": "string", "description": "素材内容（base64 或 data URL），最大 10MB" },
                                    "desc": { "type": "string", "description": "素材描述" },
                                    "source": { "type": "string", "description": "素材来源" },
                                    "tags": { "type": "string", "description": "标签（逗号分隔）" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "添加成功" },
                    "400": { "description": "素材名称和内容不能为空" },
                    "413": { "description": "素材内容过大，最大允许 10MB" }
                }
            }
        },
        "/api/asset-library/{id}": {
            "put": {
                "tags": ["asset-library"],
                "summary": "更新素材库素材",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "desc": { "type": "string" },
                                    "tags": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "更新成功" },
                    "404": { "description": "素材不存在" }
                }
            },
            "delete": {
                "tags": ["asset-library"],
                "summary": "删除素材库素材",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "删除成功" } }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        }
    },
    "security": [{ "cookieAuth": [] }]
}